requirements.txt
allure-pytest==2.8.10
pytest==5.3.5
flaky==3.6.1
requests==2.23.0